+++ frontend/packages/ux-editor/src/components/config/ExpressionContent/Expression/useExpressionTexts.ts
